Add tests documenting current ripgrep behavior
This adds (currently passing) tests expalining current interactions
between --glob, --type, --type-not ignored files, and hidden files.

This also makes the `assert_exit_code` test utility print stdout
and stderr when it fails.

The tests marked as `_legacy` are the behaviors I think should
be changed, either as part of the "globs respect ignores" proposal
or the proposal to keep track of argument order and make `--type`
an alias for `--glob`.