add complection for built-in `--hyperlink-format`s
The goal is to make the completion for `rg --hyperlink-format v<TAB>`
work in the fish shell.

These are not exhaustive (the user can also specify custom formats). If
this ever makes a difference, perhaps `fn
doc_choices_are_exhaustive(&self) -> bool` can be added to the `Flags`
trait.

The `grep+` value necessitated a change to a test.

I'm not sure whether there's a good way to generate the choices directly
from `hyperlink_aliases.rs` as a `&'static [&'static str]`. The simplest
would be to reorganize `hyperlink_aliases.rs` to keep the keys and
values in separate arrays, but that would be less readable. OTOH, the
values already need to be kept in sync with the long-form description
inside the help text, which currently also has to be an &'static str.